v2/requirements.txt
a2a-sdk[http-server]>=0.3.20
openai>=1.50.0
rank-bm25>=0.2.2
httpx>=0.28.1
pydantic>=2.11.0
uvicorn>=0.38.0
python-dotenv>=1.1.1
